fix(rate_of_closure, sidekick): hide sidekick dock initially and modularize sidebar integration helper
