+++ src/shared/python/gui_launcher/tools_sidebar_integration.py
@@ -22,13 +22,22 @@ old behaviour (return ``installed=False`` with no widget).
 
 from __future__ import annotations
 
-import importlib
-import inspect
 import logging
 from collections.abc import Callable
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
+
+from ._sidebar_integration_helpers import (
+    add_dock_widget,
+    call_shared_installer,
+    connect_file_open_request,
+    create_sidebar_from_module,
+    ensure_dock_widget,
+    get_sidekick_tokens,
+    import_sidebar_module,
+    status_sidebar,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -73,13 +82,13 @@ class NullToolsSidebar:
             # Only attempt to create a QLabel when a QApplication is running;
             # otherwise the call segfaults or raises in headless / test contexts.
             if QApplication.instance() is not None:
-                label = QLabel(
-                    text=(
-                        "<b>Sidekick Tools</b><br/>"
-                        "<small>Install the Tools workspace to enable the full sidebar.<br/>"
-                        "Run: <code>scripts/setup_tools_workspace.sh</code></small>"
-                    ),
+                notice = (
+                    "<b>Sidekick Tools</b><br/>"
+                    "<small>Install the Tools workspace to enable "
+                    "the full sidebar.<br/>"
+                    "Run: <code>scripts/setup_tools_workspace.sh</code></small>"
                 )
+                label = QLabel(text=notice)
                 label.setWordWrap(True)
                 label.setContentsMargins(12, 12, 12, 12)
                 self._widget = label
@@ -98,31 +107,6 @@ class NullToolsSidebar:
     def widget(self) -> Any:
         """Return the inner display widget."""
         return self._widget
-
-
-_SIDEBAR_MODULE_CANDIDATES = (
-    "sidekick.ui.tools_sidebar",
-    "sidekick.ui.tools_sidebar",
-    "shared.python.sidekick.ui.tools_sidebar",
-)
-_SIDEBAR_CLASS_CANDIDATES = (
-    "ToolsSidebar",
-    "UnifiedToolsSidebar",
-    "ToolsSidebarWidget",
-)
-_FILE_OPEN_SIGNAL_CANDIDATES = (
-    "file_open_requested",
-    "open_file_requested",
-    "fileRequested",
-    "openRequested",
-)
-_FILE_OPEN_METHOD_CANDIDATES = (
-    "open_file",
-    "load_file",
-    "_open_file",
-    "_load_file",
-    "load_data_file",
-)
 
 
 @dataclass(frozen=True)
@@ -170,7 +154,7 @@ def install_tools_sidebar(
             "main_window does not support dock widgets",
         )
 
-    module = _import_sidebar_module()
+    module = import_sidebar_module()
     if module is None:
         if install_fallback:
             return _install_null_sidebar(main_window)
@@ -199,20 +183,12 @@ def install_tools_sidebar(
 
 
 def _install_null_sidebar(main_window: Any) -> ToolsSidebarInstallStatus:
-    """Install a :class:`NullToolsSidebar` placeholder into *main_window*.
-
-    Args:
-        main_window: QMainWindow-like host.
-
-    Returns:
-        A :class:`ToolsSidebarInstallStatus` with ``installed=True`` and
-        ``sidebar`` pointing at the :class:`NullToolsSidebar` instance.
-    """
-    tokens = _get_sidekick_tokens()
+    """Install a :class:`NullToolsSidebar` placeholder into *main_window*."""
+    tokens = get_sidekick_tokens()
     sidebar = NullToolsSidebar(sidekick_tokens=tokens)
     try:
-        dock = _ensure_dock_widget(sidebar, main_window)
-        _add_dock_widget(main_window, dock)
+        dock = ensure_dock_widget(sidebar, main_window)
+        add_dock_widget(main_window, dock)
     except (TypeError, RuntimeError) as exc:
         # Qt not available or main_window is a non-Qt stub — still report
         # installed=True so callers know the fallback path was reached.
@@ -228,52 +204,14 @@ def _install_null_sidebar(main_window: Any) -> ToolsSidebarInstallStatus:
     )
 
 
-def _import_sidebar_module() -> Any | None:
-    from pathlib import Path
-
-    vendor_root = Path(__file__).resolve().parent.parent.parent.parent
-
-    # Try checking out sibling Tools repository first, so user's active local development in Tools takes priority
-    # sys.path mutations removed by issue #5873. We now rely on explicit shims (e.g., sidekick).
-    sibling_tools = vendor_root.parent.parent / "Tools"
-    if sibling_tools.is_dir():
-        sibling_src = str(sibling_tools / "src")  # noqa: F841
-        sibling_python = str(sibling_tools / "src" / "shared" / "python")  # noqa: F841
-
-    # Fall back to vendored ud-tools
-    # sys.path mutations removed by issue #5873. We now rely on explicit shims (e.g., sidekick).
-    vendor_src_path = str(vendor_root / "vendor" / "ud-tools" / "src")  # noqa: F841
-    vendor_python_path = str(  # noqa: F841
-        vendor_root / "vendor" / "ud-tools" / "src" / "shared" / "python"
-    )
-
-    for module_name in _SIDEBAR_MODULE_CANDIDATES:
-        try:
-            return importlib.import_module(module_name)
-        except ImportError:
-            continue
-    return None
-
-
 def is_tools_sidebar_available() -> bool:
-    """Return ``True`` when one of the shared Tools sidebar modules imports.
-
-    This is a lightweight probe intended for diagnostics: it attempts to import
-    the same candidate module names that :func:`install_tools_sidebar` would
-    try, and returns ``True`` on the first success. No widgets are created and
-    no side effects are produced beyond what ``importlib.import_module``
-    performs.
-
-    Returns:
-        ``True`` if the sibling Tools repo's sidebar module is importable in
-        the current environment, ``False`` otherwise.
-    """
-    return _import_sidebar_module() is not None
+    """Return ``True`` when one of the shared Tools sidebar modules imports."""
+    return import_sidebar_module() is not None
 
 
 def _resolved_sidebar_module_name() -> str | None:
     """Return the dotted module name of the shared sidebar, if importable."""
-    module = _import_sidebar_module()
+    module = import_sidebar_module()
     if module is None:
         return None
     name = getattr(module, "__name__", None)
@@ -291,7 +229,7 @@ def _install_from_module(
 
     module_installer = getattr(module, "install_tools_sidebar", None)
     if callable(module_installer):
-        installed = _call_shared_installer(
+        installed = call_shared_installer(
             module_installer,
             main_window=main_window,
             project_root=project_root,
@@ -305,7 +243,7 @@ def _install_from_module(
                 "shared module installer did not install sidebar",
                 module_name=module_name,
             )
-        sidebar = _status_sidebar(installed)
+        sidebar = status_sidebar(installed)
         dock = (
             getattr(installed, "dock_widget", None)
             or getattr(installed, "dock", None)
@@ -317,10 +255,10 @@ def _install_from_module(
             dock=dock,
             sidebar=sidebar,
             module_name=module_name,
-            file_open_connected=_connect_file_open_request(sidebar, main_window),
+            file_open_connected=connect_file_open_request(sidebar, main_window),
         )
 
-    sidebar = _create_sidebar_from_module(
+    sidebar = create_sidebar_from_module(
         module,
         main_window=main_window,
         project_root=project_root,
@@ -333,9 +271,9 @@ def _install_from_module(
             module_name=module_name,
         )
 
-    dock = _ensure_dock_widget(sidebar, main_window)
-    _add_dock_widget(main_window, dock)
-    connected = _connect_file_open_request(sidebar, main_window)
+    dock = ensure_dock_widget(sidebar, main_window)
+    add_dock_widget(main_window, dock)
+    connected = connect_file_open_request(sidebar, main_window)
 
     return ToolsSidebarInstallStatus(
         True,
@@ -345,272 +283,3 @@ def _install_from_module(
         module_name=module_name,
         file_open_connected=connected,
     )
-
-
-def _create_sidebar_from_module(
-    module: Any,
-    *,
-    main_window: Any,
-    project_root: Path | None,
-    context_provider: Callable[[], Any] | None,
-) -> Any | None:
-    factory = getattr(module, "create_tools_sidebar", None)
-    if callable(factory):
-        return _call_sidebar_factory(
-            factory,
-            main_window=main_window,
-            project_root=project_root,
-            context_provider=context_provider,
-        )
-
-    for class_name in _SIDEBAR_CLASS_CANDIDATES:
-        sidebar_class = getattr(module, class_name, None)
-        if callable(sidebar_class):
-            return _call_sidebar_factory(
-                sidebar_class,
-                main_window=main_window,
-                project_root=project_root,
-                context_provider=context_provider,
-            )
-    return None
-
-
-def _call_shared_installer(
-    installer: Callable[..., Any],
-    *,
-    main_window: Any,
-    project_root: Path | None,
-    context_provider: Callable[[], Any] | None,
-) -> Any:
-    kwargs: dict[str, Any] = _sidebar_factory_kwargs(
-        main_window=main_window,
-        project_root=project_root,
-        context_provider=context_provider,
-    )
-    try:
-        signature = inspect.signature(installer)
-    except (TypeError, ValueError):
-        return installer(
-            main_window, project_root=project_root, context_provider=context_provider
-        )
-
-    accepted = set(signature.parameters)
-    accepts_kwargs = any(
-        param.kind == inspect.Parameter.VAR_KEYWORD
-        for param in signature.parameters.values()
-    )
-    call_kwargs = {
-        key: value for key, value in kwargs.items() if accepts_kwargs or key in accepted
-    }
-    return installer(main_window, **call_kwargs)
-
-
-def _call_sidebar_factory(
-    factory: Callable[..., Any],
-    *,
-    main_window: Any,
-    project_root: Path | None,
-    context_provider: Callable[[], Any] | None,
-) -> Any:
-    kwargs = _sidebar_factory_kwargs(
-        main_window=main_window,
-        project_root=project_root,
-        context_provider=context_provider,
-    )
-    try:
-        signature = inspect.signature(factory)
-    except (TypeError, ValueError):
-        return factory(
-            **{key: value for key, value in kwargs.items() if value is not None}
-        )
-
-    accepted = {
-        name
-        for name, param in signature.parameters.items()
-        if param.kind
-        in (
-            inspect.Parameter.POSITIONAL_OR_KEYWORD,
-            inspect.Parameter.KEYWORD_ONLY,
-        )
-    }
-    accepts_kwargs = any(
-        param.kind == inspect.Parameter.VAR_KEYWORD
-        for param in signature.parameters.values()
-    )
-    if accepts_kwargs:
-        call_kwargs = {key: value for key, value in kwargs.items() if value is not None}
-    else:
-        call_kwargs = {
-            key: value
-            for key, value in kwargs.items()
-            if key in accepted and value is not None
-        }
-    return factory(**call_kwargs)
-
-
-def _sidebar_factory_kwargs(
-    *,
-    main_window: Any,
-    project_root: Path | None,
-    context_provider: Callable[[], Any] | None,
-) -> dict[str, Any]:
-    return {
-        "parent": main_window,
-        "project_root": project_root,
-        "context_provider": context_provider,
-        "sidekick_tokens": _get_sidekick_tokens(),
-    }
-
-
-def _get_sidekick_tokens() -> dict[str, str]:
-    try:
-        from src.shared.python.theme.sidekick_tokens import get_current_sidekick_tokens
-
-        return get_current_sidekick_tokens()
-    except Exception:  # noqa: BLE001 - sidebar startup must stay optional
-        return {}
-
-
-def _ensure_dock_widget(sidebar: Any, main_window: Any) -> Any:
-    if _looks_like_dock_widget(sidebar):
-        return sidebar
-
-    try:
-        from PyQt6.QtWidgets import QDockWidget
-    except ImportError as exc:
-        raise RuntimeError("PyQt6 is required to wrap sidebar widgets") from exc
-
-    dock = QDockWidget("Tools", main_window)
-    dock.setObjectName("unifiedToolsSidebarDock")
-    dock.setWidget(sidebar)
-    return dock
-
-
-def _looks_like_dock_widget(widget: Any) -> bool:
-    return hasattr(widget, "setWidget") and hasattr(widget, "toggleViewAction")
-
-
-def _add_dock_widget(main_window: Any, dock: Any) -> None:
-    main_window.addDockWidget(_right_dock_area(), dock)
-
-
-def _right_dock_area() -> Any:
-    try:
-        from PyQt6.QtCore import Qt
-
-        return Qt.DockWidgetArea.RightDockWidgetArea
-    except ImportError:
-        return "right"
-
-
-def _connect_file_open_request(sidebar: Any, main_window: Any) -> bool:
-    signal_owner, signal = _find_signal(sidebar)
-    if signal is None and _looks_like_dock_widget(sidebar):
-        widget = sidebar.widget() if hasattr(sidebar, "widget") else None
-        signal_owner, signal = _find_signal(widget)
-    if signal is None:
-        return False
-
-    handler = _find_file_open_handler(main_window)
-    if handler is None:
-        handler = _build_status_handler(main_window)
-
-    try:
-        signal.connect(handler)
-    except (AttributeError, TypeError) as exc:
-        logger.debug("Could not connect sidebar file-open signal: %s", exc)
-        return False
-
-    logger.debug(
-        "Connected tools sidebar file-open signal from %s",
-        type(signal_owner).__name__,
-    )
-    return True
-
-
-def _find_signal(candidate: Any) -> tuple[Any | None, Any | None]:
-    if candidate is None:
-        return None, None
-    for attr_name in _FILE_OPEN_SIGNAL_CANDIDATES:
-        signal = getattr(candidate, attr_name, None)
-        if hasattr(signal, "connect"):
-            return candidate, signal
-    return None, None
-
-
-def _find_file_open_handler(main_window: Any) -> Callable[[Any], Any] | None:
-    candidates = [main_window]
-    central_widget = _maybe_call(main_window, "centralWidget")
-    if central_widget is not None:
-        candidates.append(central_widget)
-    inner_window = _maybe_call(main_window, "inner_main_window")
-    if inner_window is not None:
-        candidates.append(inner_window)
-
-    for target in candidates:
-        for method_name in _FILE_OPEN_METHOD_CANDIDATES:
-            method = getattr(target, method_name, None)
-            if callable(method) and _callable_accepts_path(method):
-                return method
-    return None
-
-
-def _callable_accepts_path(method: Callable[..., Any]) -> bool:
-    try:
-        signature = inspect.signature(method)
-    except (TypeError, ValueError):
-        return True
-
-    positional = [
-        param
-        for param in signature.parameters.values()
-        if param.kind
-        in (
-            inspect.Parameter.POSITIONAL_ONLY,
-            inspect.Parameter.POSITIONAL_OR_KEYWORD,
-        )
-    ]
-    return any(
-        param.kind == inspect.Parameter.VAR_POSITIONAL
-        for param in signature.parameters.values()
-    ) or bool(positional)
-
-
-def _build_status_handler(main_window: Any) -> Callable[[Any], None]:
-    def _handle_requested_file(path: Any) -> None:
-        message = f"Tools sidebar requested file open: {path}"
-        status_bar = _maybe_call(main_window, "statusBar")
-        if status_bar is not None and hasattr(status_bar, "showMessage"):
-            try:
-                status_bar.showMessage(message, 5000)
-                return
-            except TypeError:
-                status_bar.showMessage(message)
-                return
-        logger.info(message)
-
-    return _handle_requested_file
-
-
-def _maybe_call(obj: Any, method_name: str) -> Any | None:
-    method = getattr(obj, method_name, None)
-    if not callable(method):
-        return None
-    try:
-        return method()
-    except TypeError:
-        return None
-
-
-def _status_sidebar(installed: Any) -> Any:
-    if installed is None:
-        return None
-    sidebar = getattr(installed, "sidebar", None)
-    if sidebar is not None:
-        return sidebar
-    if hasattr(installed, "widget") and callable(installed.widget):
-        try:
-            return installed.widget()
-        except TypeError:
-            return installed
-    return installed
